linux: Support for RNGB generator based on RNGC IMX driver (5.4.y imx6ul{l})
The imx6ull only has the RNGB random number generator (which is not
part of CAAM) which works as a standalone IP block (in conjunction
with DCP).

As imx6ull doesn't have CAAM, the RNGC iMX driver can be used to
control RNGB random number HW.

This change enables usage of RNGC driver with imx6ull RNGB HW.
Now it is possible to use the /dev/hwrng device (with e.g. cat /dev/hwrng)

(Those patches are not required on LTS 5.10.y onwards)

Signed-off-by: Lukasz Majewski <lukma@denx.de>